pyproject.toml
[project]
name = "weca-core-data"
version = "0.1.0"
description = "ETL for WECA Core Data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "dlt[duckdb,rest-api]>=1.18.2",
    "duckdb>=1.1.3",
    "fastexcel>=0.12.1",
    "geopandas>=1.0.1",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "mypy>=1.15.0",
    "pandas>=2.2.3",
    "polars>=1.20.0",
    "pyarrow>=18.1.0",
    "pyjanitor>=0.30.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.2",
    "visidata>=3.1.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.9.1",
]


